+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.1
+* added `BackupManager::delete()` method;
 * improved tests. Also errors in the shell are checked.
 
 ### 1.0.0